Moved MArrayCoordinate and MCoordinate into a single file. Also added an enum for message types between instances and instance manager.
